+++ plugin/src/gatsby-node.ts
@@ -1,0 +1,6 @@
+export type { IPluginOptions } from "./types";
+
+export { pluginOptionsSchema } from "./plugin-options-schema";
+export { sourceNodes } from "./source-nodes";
+export { createResolvers } from "./generate-image-source";
+export { createSchemaCustomization } from "./create-schema-customization";