+++ plugin/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-06-05
+
+This is a major version release, which includes several improvements, but also several
+breaking changes
+
+### Changed
+- The `photo_id` field is now simply called `_id`
+- Images and Thumbnails are now split into separate lists
+- Images are no longer exposed in an object with keys for pixel sizes, but in a list
+
+### Added
+- Support is added for gatsby-plugin-image. This is not full support, but does support the
+  major features like responsive image sizing and blur-up placeholders. 
+
 ## [1.2.0] - 2021-04-19
 
 ### Added